refactor(schema): Update edges and fields in schemas

+++ pkg/cep_ent/schema/model_price.go
@@ -15,7 +15,7 @@ type ModelPrice struct {
 // Fields of the ModelPrice.
 func (ModelPrice) Fields() []ent.Field {
 	return []ent.Field{
-		field.Int("model_id").StructTag(`json:"model_id,string"`).Comment("模型ID"),
+		field.Int64("model_id").StructTag(`json:"model_id,string"`).Comment("模型ID"),
 		field.Enum("invoke_type").Default(string(enums.UnKnownInvokeType)).GoType(enums.ApiInvokeType).StructTag(`json:"invoke_type"`).Comment("调用类型"),
 		field.Enum("gpu_version").Default(string(enums.GpuVersionUnknown)).GoType(enums.GpuVersion3060).StructTag(`json:"gpu_version"`).Comment("GPU 版本"),
 		field.Int("input_gpu_price").Default(0).StructTag(`json:"input_gpu_price"`).Comment("输入算力价格"),
